requirements.txt
article==0.1.1
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
decorator==5.1.0
Django==3.2.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-model-utils==4.2.0
django-mptt==0.13.4
django-recaptcha==2.0.6
django-render-partial==0.4
django-sslserver==0.22
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
order==1.3.4
Pillow==8.4.0
psycopg2==2.9.3
pytz==2021.3
requests==2.26.0
result==0.7.0
scinum==1.2.0
self==2020.12.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
validate-email==1.3
